add state to lazyList

+++ feature/creatorPost/src/main/java/su/afk/kemonos/creatorPost/presenter/util/IntStateMapSaver.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.saveable.mapSaver
 
 /** пересчет блока content */
 internal val IntStateMapSaver = mapSaver(
-    save = { stateMap -> stateMap.mapValues { it.value } },
+    save = { stateMap -> stateMap.mapValues { it.value.intValue } },
     restore = { restored ->
         restored.mapValues { mutableIntStateOf(it.value as Int) }.toMutableMap()
     }
